Updating weekly nutrient detail view
