+++ MC3/Views/Components/FoodList/FoodCardViewComponent.swift
@@ -1,72 +1,47 @@
 //
-//  FoodEatenCardComponentView.swift
+//  FoodCardViewComponent.swift
 //  MC3
 //
-//  Created by Jovanna Melissa on 21/08/24.
+//  Created by Jovanna Melissa on 18/08/24.
 //
 
 import SwiftUI
 
-struct FoodEatenCardComponentView: View {
+struct FoodCardViewComponent: View {
     var food:Food
-    var intakeAmount:Int
     @State private var foodStatusColor:Color = .greensuccess
     @State private var foodStatusString:String = ""
+    @State var showingSheet = false
+    @ObservedObject var vm:FoodLogViewModel
     
     var body: some View {
         ZStack{
             RoundedRectangle(cornerRadius: 24)
-                .foregroundStyle(.verylightgraybg)
+                .foregroundStyle(.bluebg)
             HStack{
                 VStack(alignment: .leading){
                     HStack{
                         Image(systemName: "circle.fill")
                             .font(.caption)
                         Text(foodStatusString)
-                        
+                            .font(.custom("Lato-Regular", size: 15))
                         Spacer()
-                        
-                        Text("\(food.calories * intakeAmount) Kcal")
-                            .fontWeight(.bold)
-                            .foregroundStyle(.darkblueprimary)
                     }
                     .foregroundStyle(foodStatusColor)
                     
-                    HStack{
-                        Text(food.title)
-                            .fontWeight(.bold)
-                            .font(.headline)
-                        
-                        Spacer()
-                        
-                        Text("\(intakeAmount) serving")
-                            .foregroundStyle(.darkgraytext)
-                    }
+                    Text(food.title)
+                        .font(.custom("Lato-Bold", size: 17))
                     
-                    ZStack{
-                        RoundedRectangle(cornerRadius: 12)
-                            .foregroundStyle(.yellowbg)
-                        
-                        HStack{
-                            Text("**\(food.protein * Float(intakeAmount), specifier: "%.0f")g** Protein")
-                            Divider()
-                            
-                            Text("**\(food.folate * Float(intakeAmount), specifier: "%.0f")mg** Folic Acid")
-                            Divider()
-                            
-                            Text("**\(food.calcium * Float(intakeAmount), specifier: "%.0f")mg** Calcium")
-                        }
-                        .foregroundStyle(.darkorangewarning)
-                        .padding(.horizontal)
-                        .font(.system(size: 13))
-                    }
-                    .frame(height: 30)
+                    Text("\(food.calories) kcal | \(food.servingSize)gr/serving")
+                        .font(.custom("Lato-Regular", size: 13))
+                        .foregroundStyle(.darkgraytext)
                     
                     NavigationLink{
-                        
+                        #warning("Panggil food detail page disini")
                     } label: {
                         HStack{
                             Text("See details")
+                                .font(.custom("Lato-Regular", size: 17))
                                 .foregroundStyle(.darkblueprimary)
                             Image(systemName: "chevron.right")
                                 .foregroundStyle(.darkblueprimary)
@@ -75,17 +50,32 @@ struct FoodEatenCardComponentView: View {
                     }
                     .padding(.top, 10)
                 }
+                
+                Button(action: {
+                    showingSheet = true
+                }, label: {
+                    Image(systemName: "plus.circle.fill")
+                        .foregroundStyle(.blueprimary)
+                        .font(.largeTitle)
+                })
             }
             .padding()
+            
         }
+        .padding()
         .frame(minHeight: 150, maxHeight: 200)
+        .sheet(isPresented: $showingSheet){
+            FoodInputSheetComponentView(food: food, showingSheet: $showingSheet, vm: vm)
+                .presentationDetents([.medium])
+                .presentationDragIndicator(.visible)
+        }
         .onAppear(perform: {
             determineStatusColor(foodStatus: food.edibleStatus.rawValue)
         })
     }
 }
 
-extension FoodEatenCardComponentView{
+extension FoodCardViewComponent{
     func determineStatusColor(foodStatus: String){
         if food.edibleStatus.rawValue == "safe"{
             foodStatusColor = .greensuccess
@@ -101,5 +91,5 @@ extension FoodEatenCardComponentView{
 }
 
 //#Preview {
-//    FoodEatenCardComponentView()
+//    FoodCardViewComponent(foodName: "Rujak Cingur Vegetarian Khas Surabaya Barat Daya", foodStatus: "safe", foodCalorie: 257, foodServing: 300, foodProtein: 39, foodFolicAcid: 39, foodCalcium: 39)
 //}
